// src/app/shared/Search-table-heder/model/sheaech-table.ts
export class SearchTable {
  inputTextName: string = 'inputText1';
  lapel: string = 'shared.searchTableBar.hintFieldFilter1';
  hintInput: string = 'shared.searchTableBar.hintFieldFilter1';
  IsAutoComplete: boolean = false;
  IsDropDawn: boolean = false;
  IsInput: boolean = true;
  item: any[] = [];
  position: string = '1';
}